+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Version 0.10.0.9000
 
+## mini.hues
+
+- BREAKING: Stop supporting deprecated 'HiPhish/nvim-ts-rainbow2'.
+
 
 # Version 0.10.0